GH-01: MVC-JPA-Thymeleaf - Part 05
