+++ .gitignore
@@ -2,5 +2,6 @@
 **/.project
 **/.settings
 **/.classpath
+**/.factorypath
 **/target
