+++ 01-mvc-jpa-thymeleaf/final/pom.xml
@@ -8,11 +8,10 @@
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
     <version>2.5.0</version>
-    <relativePath />
-    <!-- lookup parent from repository -->
+    <relativePath /><!-- lookup parent from repository -->
   </parent>
 
-  <groupId>io.codejournal.springprojects</groupId>
+  <groupId>io.codejournal.springboot</groupId>
   <artifactId>mvc-jpa-thymeleaf</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <name>mvc-jpa-thymeleaf</name>
@@ -46,12 +45,6 @@
       <artifactId>spring-boot-starter-security</artifactId>
     </dependency>
 
-    <!-- View -->
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-thymeleaf</artifactId>
-    </dependency>
-
     <!-- Database -->
     <dependency>
       <groupId>org.springframework.boot</groupId>
@@ -61,6 +54,14 @@
       <groupId>com.h2database</groupId>
       <artifactId>h2</artifactId>
     </dependency>
+
+    <!-- View -->
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-thymeleaf</artifactId>
+    </dependency>
+
+
 
     <!-- Testing -->
     <dependency>
