+++ 01-mvc-jpa-thymeleaf/final/src/main/java/io/codejournal/springboot/mvcjpathymeleaf/entity/User.java
@@ -1,0 +1,28 @@
+package io.codejournal.springboot.mvcjpathymeleaf.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "user")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class User {
+
+    @Id
+    @Column(name = "u_username")
+    private String username;
+
+    @Column(name = "u_password")
+    private String password;
+
+    @Column(name = "u_role")
+    private String role;
+}
