+++ 01-mvc-jpa-thymeleaf/final/src/main/java/io/codejournal/springboot/mvcjpathymeleaf/repository/UserRepository.java
@@ -1,0 +1,14 @@
+package io.codejournal.springboot.mvcjpathymeleaf.repository;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import io.codejournal.springboot.mvcjpathymeleaf.entity.User;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+
+    Optional<User> findByUsername(final String username);
+}
