--- 01-mvc-jpa-thymeleaf/final/src/main/java/io/codejournal/springprojects/mvcjpathymeleaf/entity/Student.java
@@ -1,32 +1,0 @@
-package io.codejournal.springprojects.mvcjpathymeleaf.entity;
-
-import java.util.UUID;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "student")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class Student {
-
-    @Id
-    @GeneratedValue(generator = "UUID")
-    @Column(name = "s_id")
-    private UUID id;
-
-    @Column(name = "first_name")
-    private String firstName;
-
-    @Column(name = "last_name")
-    private String lastName;
-}
