--- 01-mvc-jpa-thymeleaf/final/src/main/java/io/codejournal/springprojects/mvcjpathymeleaf/service/UserService.java
@@ -1,50 +1,0 @@
-package io.codejournal.springprojects.mvcjpathymeleaf.service;
-
-import static java.util.Collections.singletonList;
-
-import java.util.Optional;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import io.codejournal.springprojects.mvcjpathymeleaf.entity.User;
-import io.codejournal.springprojects.mvcjpathymeleaf.repository.UserRepository;
-
-@Service
-public class UserService implements UserDetailsService {
-
-    private final UserRepository repository;
-
-    @Autowired
-    public UserService(final UserRepository repository) {
-        this.repository = repository;
-    }
-
-    public User save(final User user) {
-        return repository.save(user);
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(final String username) throws UsernameNotFoundException {
-
-        final Optional<User> entry = repository.findByUsername(username);
-
-        if (entry.isEmpty()) {
-            throw new UsernameNotFoundException("User not found with Name - " + username);
-        }
-
-        final User user = entry.get();
-
-        // @formatter:off
-        return new org.springframework.security.core.userdetails.User(
-                        user.getUsername(),
-                        user.getPassword(),
-                        singletonList(new SimpleGrantedAuthority(user.getRole()))
-        );
-        // @formatter:on
-    }
-}
