+++ 01-mvc-jpa-thymeleaf/final/src/test/java/io/codejournal/springboot/mvcjpathymeleaf/ApplicationTests.java
@@ -1,0 +1,13 @@
+package io.codejournal.springboot.mvcjpathymeleaf;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
