--- 01-mvc-jpa-thymeleaf/final/src/test/java/io/codejournal/springprojects/mvcjpathymeleaf/ApplicationTests.java
@@ -1,13 +1,0 @@
-package io.codejournal.springprojects.mvcjpathymeleaf;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
