--- 01-mvc-jpa-thymeleaf/final/src/test/java/io/codejournal/springprojects/mvcjpathymeleaf/repository/UserRepositoryTest.java
@@ -1,42 +1,0 @@
-package io.codejournal.springprojects.mvcjpathymeleaf.repository;
-
-import static java.util.UUID.randomUUID;
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Optional;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import io.codejournal.springprojects.mvcjpathymeleaf.entity.User;
-
-@ExtendWith(SpringExtension.class)
-@DataJpaTest
-public class UserRepositoryTest {
-
-    @Autowired
-    private UserRepository repository;
-
-    @Autowired
-    private TestEntityManager entityManager;
-
-    @Test
-    public void findByUsername_ReturnsUser_WhenRecordExists() {
-
-        final String username = randomUUID().toString();
-
-        final User expected = new User(username, randomUUID().toString(), randomUUID().toString());
-
-        entityManager.persist(expected);
-        entityManager.flush();
-
-        final Optional<User> actual = repository.findByUsername(username);
-
-        assertThat(actual.isPresent()).isTrue();
-        assertThat(actual.get()).isEqualTo(expected);
-    }
-}
