+++ 01-mvc-jpa-thymeleaf/part02/pom.xml
@@ -7,7 +7,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.5.0</version>
+    <version>2.6.4</version>
     <relativePath /> <!-- lookup parent from repository -->
   </parent>
