+++ 01-mvc-jpa-thymeleaf/part04/pom.xml
@@ -7,7 +7,7 @@
   <parent>
     <groupId>org.springframework.boot</groupId>
     <artifactId>spring-boot-starter-parent</artifactId>
-    <version>2.5.0</version>
+    <version>2.6.4</version>
     <relativePath /> <!-- lookup parent from repository -->
   </parent>
 
@@ -15,7 +15,7 @@
   <artifactId>mvc-jpa-thymeleaf</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <name>mvc-jpa-thymeleaf</name>
-  <description>Demo project for Spring Boot</description>
+  <description>Project to demostrate Spring MVC, JPA and Thymeleaf</description>
 
   <properties>
     <java.version>11</java.version>
