+++ 01-mvc-jpa-thymeleaf/part05/pom.xml
@@ -15,7 +15,7 @@
   <artifactId>mvc-jpa-thymeleaf</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <name>mvc-jpa-thymeleaf</name>
-  <description>Demo project for Spring Boot</description>
+  <description>Project to demostrate Spring MVC, JPA and Thymeleaf</description>
 
   <properties>
     <java.version>11</java.version>
@@ -39,6 +39,10 @@
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-undertow</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-security</artifactId>
     </dependency>
 
     <!-- Database -->
