+++ 01-mvc-jpa-thymeleaf/part05/src/main/java/io/codejournal/springboot/mvcjpathymeleaf/entity/Student.java
@@ -1,0 +1,32 @@
+package io.codejournal.springboot.mvcjpathymeleaf.entity;
+
+import java.util.UUID;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "student")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Student {
+
+    @Id
+    @GeneratedValue(generator = "UUID")
+    @Column(name = "s_id")
+    private UUID id;
+
+    @Column(name = "s_first_name")
+    private String firstName;
+
+    @Column(name = "s_last_name")
+    private String lastName;
+}
