+++ 01-mvc-jpa-thymeleaf/part05/src/main/java/io/codejournal/springboot/mvcjpathymeleaf/entity/User.java
@@ -1,0 +1,25 @@
+package io.codejournal.springboot.mvcjpathymeleaf.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "user")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class User {
+
+    @Id
+    @Column(name = "u_username")
+    private String username;
+
+    @Column(name = "u_password")
+    private String password;
+}
