+++ 01-mvc-jpa-thymeleaf/part05/src/main/resources/templates/students/header.html
@@ -1,0 +1,14 @@
+<nav class="navbar navbar-expand-lg navbar-light bg-light" th:fragment="header">
+    <div class="container-fluid">
+        <div class="navbar-header">
+            <h1 class="navbar-brand">Student CRUD Operations</h1>
+        </div>
+        <div class="d-flex">
+            <form action="/logout" method="POST">
+                <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}" />
+                <input type="submit" class="btn btn-primary" value="Logout" />
+            </form>
+        </div>
+    </div>
+    <br/>
+</nav>
