+++ 01-mvc-jpa-thymeleaf/part05/src/main/resources/templates/students/list.html
@@ -1,0 +1,79 @@
+<!DOCTYPE html>
+
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+
+<head>
+    <title>Student List</title>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <script src="http://localhost:35729/livereload.js"></script>
+    <style>
+        @import url('https://fonts.googleapis.com/css2?family=Rubik:wght@300&display=swap');
+
+        body {
+            font-family: 'Rubik', sans-serif;
+        }
+    </style>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-+0n0xVW2eSR5OomGNYDnhzAbDsOXxcvSN1TPprVMTNDbiYZCxYbOOl7+AMvyTG2x" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-gtEjrD/SeCtmISkJkNUaaKMoLD0//ElJ19smozuHV6z3Iehds+3Ulb9Bn9Plx0x4"
+        crossorigin="anonymous"></script>
+</head>
+
+<body>
+    <nav th:replace="students/header :: header"></nav>
+
+    <br />
+    <br />
+
+    <div class="container-sm">
+        <table class="table table-bordered table-striped table-hover">
+            <thead>
+                <tr>
+                    <th style="width: 10%; text-align: center">#</th>
+                    <th style="width: 30%;">First Name</th>
+                    <th style="width: 30%;">Last Name</th>
+                    <th style="width: 30%;">&nbsp;</th>
+                </tr>
+            </thead>
+            <tbody>
+                <tr th:each="student,iteration : ${students}">
+                    <td th:text="${(currentPageNumber * pageSize) + iteration.count}">1</td>
+                    <td th:text="${student.firstName}">John</td>
+                    <td th:text="${student.lastName}">Doe</td>
+                    <td>
+                        <a class="btn btn-info" th:href="'view?id=' + ${student.id}">View</a>
+                        <a class="btn btn-warning" th:href="'edit?id=' + ${student.id}">Edit</a>
+                        <a class="btn btn-danger" th:href="'delete?id=' + ${student.id}">Delete</a>
+                    </td>
+                </tr>
+            </tbody>
+        </table>
+
+        <nav class="navbar navbar-expand-lg navbar-light">
+            <div class="container-fluid">
+                <a href="add" class="btn btn-primary">Add new</a>
+                <div class="d-flex">
+                    <nav aria-label="Page navigation example">
+                        <ul class="pagination">
+                            <li class="page-item disabled" th:if="${previousPageNumber == -1}">
+                                <a class="page-link" href="#">Previous</a>
+                            </li>
+                            <li class="page-item" th:if="${previousPageNumber != -1}">
+                                <a class="page-link" th:href="'?page=' + ${previousPageNumber}">Previous</a>
+                            </li>
+                            <li class="page-item disabled" th:if="${nextPageNumber == -1}">
+                                <a class="page-link" href="#">Next</a>
+                            </li>
+                            <li class="page-item" th:if="${nextPageNumber != -1}">
+                                <a class="page-link" th:href="'?page=' + ${nextPageNumber}">Next</a>
+                            </li>
+                        </ul>
+                    </nav>
+                </div>
+            </div>
+        </nav>
+    </div>
+</body>
+
+</html>
