+++ 01-mvc-jpa-thymeleaf/part05/src/test/java/io/codejournal/springboot/mvcjpathymeleaf/MvcJpaThymeleafApplicationTests.java
@@ -1,0 +1,13 @@
+package io.codejournal.springboot.mvcjpathymeleaf;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MvcJpaThymeleafApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
